@@ -18,10 +18,11 @@ setuptools.setup(
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ],
-    python_requires='>=3.6',
-    install_requires=['pynput       >=1.6.8',
+    python_requires='>=3.8',
+    install_requires=['pynput       >=1.7.1',
                       'PyAudio      >=0.2.11',
-                      'numpy        >=1.18.1',
-                      'scipy        >=1.4.1',
-                      'matplotlib   >=3.2.1'],
+                      'numpy        >=1.19.1',
+                      'scipy        >=1.5.1',
+                      'matplotlib   >=3.3.1',
+                      'scikit-image >=0.17.2'],
 )
