@@ -14,7 +14,7 @@ to enhance special non-equal tunings yourself) and a given pitch value
 (input value). The deviation in units of cent is shown, too low (red),
 too high (green).
 
-Inharmonicity of strings is considered by the equation
+Inharmonicity of strings is considered through equation
 f_n = n * f_1 * sqrt(1 + B * n**2), where n = 1, 2, 3, ... and
 B is the inharmonicity coefficient. The maximum inharmonicity accepted is
 defined in parameters.py Change accordingly for harpsichords and pianos.
@@ -28,8 +28,9 @@ NUMBER 1 JANUARY 1964
 (MA of Science), Aalto University, School of Electrical Engineering (2019)
 
 The hotkeys ctrl-y and ctrl-x exits and stops the program, respectively,
-ESC to resume. Ctrl-j and ctrl-k shorten and lengthen the recording interval,
-whereas ctrl-n and ctrl-m diminish and increase the max frequency displayed.
+ESC to resume. Ctrl-j and ctrl-k shorten and lengthen the shift between the
+individual audio slices, whereas ctrl-n and ctrl-m diminish and increase the
+max frequency displayed. Parameters can be adjusted in the parameters.py file.
 
 This program comes with ABSOLUTELY NO WARRANTY.
 This is free software, and you are welcome to redistribute it under
@@ -40,6 +41,7 @@ import pyaudio
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.collections import EventCollection
+from skimage import util
 import timeit
 import time
 from pynput import keyboard
@@ -53,13 +55,12 @@ __author__ = "Dr. Ralf Antonius Timmermann"
 __copyright__ = "Copyright (C) Dr. Ralf Antonius Timmermann"
 __credits__ = ""
 __license__ = "GPLv3"
-__version__ = "1.0.1"
+__version__ = "2.0.0"
 __maintainer__ = "Dr. Ralf A. Timmermann"
 __email__ = "rtimmermann@astro.uni-bonn.de"
 __status__ = "QA"
 
 print(__doc__)
-
 
 logging.basicConfig(format=parameters.myformat,
                     level=logging.INFO,
@@ -69,22 +70,20 @@ if parameters.DEBUG:
 
 
 class Tuner:
-
-    def __init__(self, tuning, a1):
+    def __init__(self, **kwargs):
         """
         :param a1: float
             pitch frequency for a1
         :param tuning: string
             tuning temperament
         """
-        self.record_seconds = 2.  # lengths of audio signal chunks in 1st subplot, can be adjusted by hotkey
-        self.fmax = 2000.  # maximum frequency display in 2nd subplot, can be adjusted by hotkey
-        self.a1 = a1
-        self.tuning = tuning  # see tuningTable.py
+        self.step = parameters.SLICE_SHIFT
+        self.fmax = parameters.FREQUENCY_MAX
+        self.a1 = kwargs.get('a1')
+        self.tuning = kwargs.get('tuning')  # see tuningTable.py
         self.rc = None
-        self.visible = True
 
-        self.callback_output = []
+        self.callback_output = list()
         audio = pyaudio.PyAudio()
         self.stream = audio.open(format=pyaudio.paInt16,
                                  channels=1,
@@ -108,47 +107,41 @@ class Tuner:
 
     def on_activate_x(self):
         # suspend the audio stram and freezes the plat
-        print("continue with ESC")
+        print("continue with ESC or quit with ctrl-y")
         self.rc = 'x'
 
     def on_activate_y(self):
         # exits the program
-        if self.stream.is_active(): 
-            self.stream.stop_stream()
         print("quitting...")
         self.rc = 'y'
 
     def on_activate_k(self):
-        # increases the sampling time by .1 s
-        self.record_seconds += 0.1
-        if self.record_seconds > 5.0: 
-            self.record_seconds = 5.0
-        print("Recording Time: {0:1.1f}s".format(self.record_seconds))
+        # increases the shift by which the slices progress
+        self.step += 1024
+        if self.step > parameters.SLICE_LENGTH:
+            self.step = parameters.SLICE_LENGTH
+        print("Slice shift: {0:d} bytes".format(self.step))
 
     def on_activate_j(self):
-        # decreases the sampling time by .1 s
-        self.record_seconds -= 0.1
-        if self.record_seconds < 0.5: 
-            self.record_seconds = 0.5
-        print("Recording Time: {0:1.1f}s".format(self.record_seconds))
+        # decreases the shift by which the slices progress
+        self.step -= 1024
+        if self.step < 4096:
+            self.step = 4096
+        print("Slice shift: {0:d} bytes".format(self.step))
 
     def on_activate_n(self):
         # decreases the max. frequency plotted
         self.fmax -= 500 if self.fmax > 2000 else 100
-        if self.fmax < 500: 
+        if self.fmax < 500:
             self.fmax = 500
         print("Max frequency displayed: {0:1.0f}Hz".format(self.fmax))
 
     def on_activate_m(self):
         # increases the max. frequency plotted
         self.fmax += 500 if self.fmax >= 2000 else 100
-        if self.fmax > 15000: 
+        if self.fmax > 15000:
             self.fmax = 15000
         print("Max frequency displayed: {0:1.0f}Hz".format(self.fmax))
-
-    def on_activate_v(self):
-        # toggles on/off the plot in the time domain
-        self.visible = not self.visible
 
     def on_activate_esc(self):
         # resume the audio streaming
@@ -165,9 +158,11 @@ class Tuner:
         float, None
             offset from true key in cent or None if None found or error
         """
+
         def timeusage():
             _stop = timeit.default_timer()
-            logging.debug("time utilized for key finding [s]: " + str(_stop - _start))
+            logging.debug("time utilized for key finding [s]: {0}".format(
+                str(_stop - _start)))
 
         _start = timeit.default_timer()
 
@@ -176,21 +171,72 @@ class Tuner:
             for key, value in tuningtable[self.tuning].items():
                 displaced = offset + tuningtable[self.tuning].get('A') - value
                 if -60 < displaced < 60:
-                    logging.debug(str(i) + " " +
-                                  str(key) + " " +
-                                  str(value) + " " +
-                                  str(displaced + tuningtable[self.tuning].get('A') - value))
+                    logging.debug("{0} {1} {2} {3}"
+                                  .format(str(i),
+                                          str(key),
+                                          str(value),
+                                          str(displaced +
+                                              tuningtable[self.tuning].get('A')
+                                              - value)))
                     timeusage()
                     return key, displaced
 
         timeusage()
         return None, None
 
+    def slice(self):
+        """
+        collects the audio data and arranges it into slices of length
+        SLICES_LENGTH shifted by SLICES_SHIFT
+        :return:
+        ndarray
+            (rolling) window view of the input array. If arr_in is
+            non-contiguous, a copy is made.
+        """
+        # interrupt on hotkey 'ctrl-x', resume on 'esc' and clear buffer
+        if self.rc == 'x':
+            self.stream.stop_stream()
+            while self.rc not in ['esc', 'y']:
+                # loop and wait until ESC or ctrl-y is pressed
+                time.sleep(.1)
+            if self.rc == 'esc':
+                self.callback_output = list()
+                self.stream.start_stream()
+                logging.info(
+                    "Clearing audio buffer and resuming audio stream ...")
+                self.rc = None
+            elif self.rc == 'y':
+                return None
+        # exit program on hotkey ctrl-y
+        elif self.rc == 'y':
+            self.stream.stop_stream()
+            return None
+
+        logging.debug("=== starting cycle ===")
+        _start = timeit.default_timer()
+        # wait until buffer filled by at least one FFT slice, where
+        # length is in units of buffer = 1024
+        while len(self.callback_output) < parameters.SLICE_LENGTH // 1024:
+            time.sleep(0.001)
+        # Convert the list of numpy-arrays into a 1D array (column-wise)
+        amp = np.hstack(self.callback_output)
+        slices = util.view_as_windows(amp,
+                                      window_shape=(
+                                          parameters.SLICE_LENGTH,),
+                                      step=self.step)
+        _stop = timeit.default_timer()
+        logging.debug("Audio shape: {0}, Sliced audio shape: {1}"
+                      .format(amp.shape,
+                              slices.shape))
+        logging.debug("time utilized for Audio [s]: {0}".format(
+            str(_stop - _start)))
+
+        return slices
+
     def animate(self):
         """
-        calling routine for audio, FFT, peak and partials and key finding, and plotting. Listens for events in plot
-        window
-
+        calling routine for audio, FFT, peak and partials and key finding,
+        and plotting. Listens for events in plot window
         :return:
         string
             return code
@@ -198,162 +244,151 @@ class Tuner:
         _firstplot = True
         plt.ion()  # Stop matplotlib windows from blocking
 
+        resolution = parameters.RATE / parameters.SLICE_LENGTH * 1.5
+        logging.debug(
+            "Resolution incl. Hanning apodization (Hz/channel) ~ {0}"
+            .format(str(resolution)))
+
         # start Recording
         self.stream.start_stream()
 
+        # main loop while audio stream active
         while self.stream.is_active():
-
-            _start = timeit.default_timer()
-            logging.info('Started Audio Stream ...')
-
-            time.sleep(self.record_seconds)
-            self.stream.stop_stream()  # stop the input stream for the time being
-            logging.info('Stopped Audio Stream ...')
-
-            # Convert the list of numpy-arrays into a 1D array (column-wise)
-            amp = np.hstack(self.callback_output)
-            # clear input stream
-            self.callback_output = []
-
-            # interrupt on hotkey 'ctrl-x' and resume on 'esc'
-            if self.rc == 'x':
-                while self.rc != 'esc':  # loop and wait until ESC ist pressed
-                    time.sleep(.1)
-                self.rc = None
-                self.stream.start_stream()
-                logging.info('Dump last audio stream ...')
-                continue
-            elif self.rc == 'y':
+            slices = self.slice()
+            if self.rc == 'y':
                 return self.rc
 
-            _stop = timeit.default_timer()
-            logging.debug("time utilized for Audio [s]: " + str(_stop - _start))
-            logging.info('Analyzing ...')
+            # work off all slices, before pulling from audio stream
+            for sl in slices:
+                logging.debug("no of slices: " + str(len(slices)))
+                # remove first step of slice
+                del self.callback_output[0:self.step // 1024]
+                # calculate FFT
+                t1, yfft = fft(amp=sl)
+                # peakfinding
+                peaks = peak(frequency=t1,
+                             spectrum=yfft)
+                peaklist = list(map(itemgetter(0), peaks))
+                # harmonics
+                if peaks is not None:
+                    f_measured = harmonics(peaks=peaks)  # find the key
 
-            samples = len(amp)
-            logging.info('Number of samples: ' + str(samples))
-            t = np.arange(samples) / parameters.RATE
-            resolution = parameters.RATE / samples
-            logging.info('Resolution (Hz/channel): ' + str(resolution))
+                displayed_text = ""
+                color = None
+                displayed_title = "{0:s} (a1={1:3.0f}Hz)".format(self.tuning,
+                                                                 self.a1)
+                info_text = "Resolution: {2:3.1f} Hz/channel\n" \
+                            "Audio shape: {0} [slices, samples]\n" \
+                            "Slice shift: {1:d} samples".format(slices.shape,
+                                                                self.step,
+                                                                resolution)
+                info_color = 'red' if slices.shape[0] > 3 else 'black'
+                font_title = {'family': 'serif',
+                              'color': 'darkred',
+                              'weight': 'normal',
+                              'size': 14}
 
-            # calculate FFT
-            t1, yfft = fft(amp=amp)
+                if len(f_measured) != 0:
+                    # if key is found print it and its offset colored red/green
+                    tone, displaced = self.find(f_measured=f_measured[0])
+                    if tone:
+                        displayed_text = \
+                            "{0:s} offset={1:.0f} cent".format(tone, displaced)
+                        color = 'green' if displaced >= 0 else 'red'
 
-            # peakfinding
-            peaks = peak(frequency=t1,
-                         spectrum=yfft)
-            peaklist = list(map(itemgetter(0), peaks))
+                """Matplotlib block"""
+                _start = timeit.default_timer()
+                if _firstplot:
+                    # Setup figure, axis, lines, text and initiate plot once
+                    # and copy background
+                    fig = plt.gcf()
+                    ax1 = fig.add_subplot(111)
+                    fig.set_size_inches(12, 6)
+                    ln1, = ax1.plot(t1, yfft)
+                    text = ax1.text(self.fmax, np.max(yfft), "",
+                                    verticalalignment='top',
+                                    horizontalalignment='right',
+                                    fontsize=12,
+                                    fontweight='bold'
+                                    )
+                    text1 = ax1.text(0., np.max(yfft), "",
+                                     horizontalalignment='left',
+                                     verticalalignment='top')
+                    ax1.set_title(label=displayed_title,
+                                  loc='right',
+                                  fontdict=font_title)
+                    ax1.set_xlabel('Frequency/Hz')
+                    ax1.set_ylabel('Intensity/arb. units')
+                    ax1background = fig.canvas.copy_from_bbox(ax1.bbox)
+                else:
+                    ln1.set_xdata(t1)
+                    ln1.set_ydata(yfft)
+                # set text attributes of lower subplot
+                ax1.set_xlim([0., self.fmax])
+                text.set_text(displayed_text)
+                text.set_color(color)
+                text.set_x(self.fmax)
+                text.set_y(np.max(yfft))
+                text1.set_text(info_text)
+                text1.set_color(info_color)
+                text1.set_x(0.)
+                text1.set_y(np.max(yfft))
 
-            if peaks is not None:
-                f_measured = harmonics(peaks=peaks)  # find the key
+                # remove all collections: last object first (reverse)
+                while ax1.collections:
+                    ax1.collections.pop()
+                yevents = EventCollection(positions=peaklist,
+                                          color='tab:orange',
+                                          linelength=0.05 * np.max(yfft),
+                                          linewidth=2.
+                                          )
+                ax1.add_collection(yevents)
+                yevents1 = EventCollection(positions=f_measured,
+                                           color='tab:red',
+                                           linelength=0.05 * np.max(yfft),
+                                           lineoffset=-0.04 * np.max(yfft),
+                                           linewidth=2.
+                                           )
+                ax1.add_collection(yevents1)
 
-            displayed_text = ""
-            color = 'none'
-            if len(f_measured) != 0:  # if key is found print it and its offset colored red or green
-                tone, displaced = self.find(f_measured=f_measured[0])
-                if tone:
-                    displayed_text = "{2:s} (a1={3:3.0f}Hz) {0:s} offset={1:.0f} cent"\
-                        .format(tone, displaced, self.tuning, self.a1)
-                    color = 'green' if displaced >= 0 else 'red'
+                # Rescale the axis so that the data can be seen in the plot
+                # if you know the bounds of your data you could just set this
+                # once, such that the axis don't keep changing
+                ax1.relim()
+                ax1.autoscale_view()
 
-            _start = timeit.default_timer()
-            """
-            https://stackoverflow.com/questions/40126176/fast-live-plotting-in-matplotlib-pyplot
-            """
-            if _firstplot:  # instantiate first plot and copy background
-                # Setup figure, axis, lines, text and initiate plot once and copy background
-                fig = plt.gcf()
-                ax = fig.add_subplot(211)
-                ax1 = fig.add_subplot(212)
-                fig.set_size_inches(12, 8)
-                ln, = ax.plot(t, amp)
-                ln1, = ax1.plot(t1, yfft)
-                text = ax1.text(self.fmax, np.max(yfft), "",
-                                # color='',
-                                verticalalignment='top',
-                                horizontalalignment='right',
-                                fontsize=11,
-                                fontweight='bold'
-                                )
-                ax.set_xlabel('Time/s')
-                ax.set_ylabel('Intensity/arb. units')
-                ax1.set_xlabel('Frequency/Hz')
-                ax1.set_ylabel('Intensity/arb. units')
-                axbackground = fig.canvas.copy_from_bbox(ax.bbox)
-                ax1background = fig.canvas.copy_from_bbox(ax1.bbox)
-            else:
-                # upper subplot
-                ln.set_xdata(t)
-                ln.set_ydata(amp)
-                # lower subplot
-                ln1.set_xdata(t1)
-                ln1.set_ydata(yfft)
-            ax.set_xlim([0., np.max(t)])
-            ax1.set_xlim([0., self.fmax])
-            # set text attributes of lower subplot
-            text.set_text(displayed_text)
-            text.set_color(color)
-            text.set_x(self.fmax)
-            text.set_y(np.max(yfft))
+                if _firstplot:
+                    plt.pause(0.0001)
+                    _firstplot = False
+                else:
+                    # restore background
+                    fig.canvas.restore_region(ax1background)
+                    # redraw just the points
+                    ax1.draw_artist(ln1)
+                    ax1.draw_artist(text)
+                    ax1.draw_artist(text1)
+                    # fill in the axes rectangle
+                    fig.canvas.blit(ax1.bbox)
 
-            # remove all collections: beginning from last object
-            while ax1.collections:
-                ax1.collections.pop()
-            yevents = EventCollection(positions=peaklist,
-                                      color='tab:orange',
-                                      linelength=0.05*np.max(yfft),
-                                      linewidth=2.
-                                      )
-            ax1.add_collection(yevents)
-            yevents1 = EventCollection(positions=f_measured,
-                                       color='tab:red',
-                                       linelength=0.05*np.max(yfft),
-                                       lineoffset=-0.04 * np.max(yfft),
-                                       linewidth=2.
-                                       )
-            ax1.add_collection(yevents1)
+                fig.canvas.flush_events()
+                # resume the audio streaming, expect some retardation for the
+                # status change
 
-            # Rescale the axis so that the data can be seen in the plot
-            # if you know the bounds of your data you could just set this once
-            # so that the axis don't keep changing
-            ax.relim()
-            ax.autoscale_view()
-            ax1.relim()
-            ax1.autoscale_view()
-            if _firstplot:
-                plt.pause(0.001)
-#                fig.canvas.draw()
-                _firstplot = False
-            else:
-                # restore background
-                fig.canvas.restore_region(axbackground)
-                fig.canvas.restore_region(ax1background)
-                # redraw just the points
-                ax.draw_artist(ln)
-                ax1.draw_artist(ln1)
-                ax1.draw_artist(text)
-                # fill in the axes rectangle
-                fig.canvas.blit(ax.bbox)
-                fig.canvas.blit(ax1.bbox)
-            ax.set_visible(self.visible)
-
-            fig.canvas.flush_events()
-            # resume the audio streaming, expect some retardation for the status change
-            self.stream.start_stream()
-
-            _stop = timeit.default_timer()
-            logging.debug("time utilized for matplotlib [s]: " + str(_stop - _start))
+                _stop = timeit.default_timer()
+                logging.debug("time utilized for matplotlib [s]: {0}".format(
+                    str(_stop - _start)))
 
         return self.rc
 
 
 def main():
-
     for tune in tuningtable.keys():
-        print("Temperament ({1:d}) {0:s}".format(tune, list(tuningtable).index(tune)))
+        print("Temperament ({1:d}) {0:s}".format(tune,
+                                                 list(tuningtable).index(tune)))
 
-    a = Tuner(tuning=list(tuningtable.keys())[int(input("Temperament [no]?: "))],
-              a1=float(input("A4 pitch frequency [Hz]?: ")))
+    a = Tuner(
+        tuning=list(tuningtable.keys())[int(input("Temperament [no]?: "))],
+        a1=float(input("A4 pitch frequency [Hz]?: ")))
 
     h = keyboard.GlobalHotKeys({
         '<ctrl>+x': a.on_activate_x,
@@ -362,7 +397,6 @@ def main():
         '<ctrl>+k': a.on_activate_k,
         '<ctrl>+m': a.on_activate_m,
         '<ctrl>+n': a.on_activate_n,
-        '<ctrl>+v': a.on_activate_v,
         '<esc>': a.on_activate_esc})
     h.start()
 
